[theme] move test of internal helper into internal test file
